settings refactor | api endpoint fixes on prod
